requirements.txt
python-dotenv
awscli
boto3
